Move badges logic/queries to concern
