+++ app/components/developers/badges_component.html.erb
@@ -8,7 +8,7 @@
   <% if source_contributor? %>
     <%= render BadgeComponent.new(t(".source_contributor"), color: "gray") %>
   <% end %>
-  <% if response_rate >= 90 %>
+  <% if response_rate >= HasBadges::HIGH_RESPONSE_RATE_CUTTOFF %>
     <%= render BadgeComponent.new(t(".high_response_rate"), color: "purple") %>
   <% end %>
 </div>
